requirements.txt
openai>=1.12.0
Flask>=3.0.3
gspread>=5.12.4
oauth2client>=4.1.3
python-dotenv>=1.0.1
SpeechRecognition>=3.10.0